requirements.txt
openai
composio-core
requests
python-dotenv
beautifulsoup4
